Fixed: Detect label dialog - remove label
- main.py
- 각 이미지 디텍션 라벨 생성 시 remove_label 아규먼트 값이 False면 기존 존재하는 박스 라벨들을 삭제 하지 않음
- config 설정에 따라 추론한 클래스 넘버가 클래스 넘버 변경 설정에 존재하면 변경 클래스로 진행, 없다면 오리지널 클래스로 진행
  - 어차피 바운딩 박스 라벨 필드 범위 밖의 객체는 추가되지 않음

+++ main.py
@@ -1273,8 +1273,9 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             return
 
         # Delete Current image box label
-        self.db_manager.delete_label_data(image_data_id=image_idx, is_box=1)
-        self.clear_ui_bbox_label_data()
+        if remove_label is True:
+            self.db_manager.delete_label_data(image_data_id=image_idx, is_box=1)
+            self.clear_ui_bbox_label_data()
 
         if self.detector is None:
             self.detector = ObjectDetector(cfg=self.cfg)
@@ -1289,6 +1290,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             abs_xyxy = d[:4]
             rel_xyxy = absxyxy_to_relxyxy(abs_xyxy, img_w, img_h)
             cls = int(d[5])
+            cls = self.cfg.det_infer_cls.get(cls) if self.cfg.det_infer_cls.get(cls) else cls
             if cls not in cls_idx_to_name.keys():
                 continue
             ret_num += 1
